Day 17. One of those days you kinda have to look at the input and manually do shenanigans
Debated doing a generalized solution but I have a bunch of assumptions that i dont know are actually true. For example: the jnz instruction is always last, and is always preceded by the out instruction. Or the fact that in my case I have the assumption that B & C are overwritten every loop. I **think** these are happening to everyone, but I don't wanna spend time in a generalized solution and then have that not be true.